PT 51 ST 1 Extract Configuration Component
	- CompanyNameConfig - new configuration service component for filter
	- CompanyReplaceFilter - inject configuration service instead of
	properties configuration in filter
	- change configurations for runmodes
