+++ core/src/main/java/com/aemtask/core/config/CompanyNameConfig.java
@@ -1,0 +1,6 @@
+package com.aemtask.core.config;
+
+public interface CompanyNameConfig{
+    String getSource();
+    String getReplacement();
+}
