+++ core/src/main/java/com/aemtask/core/filters/CompanyReplaceFilter.java
@@ -1,10 +1,12 @@
 package com.aemtask.core.filters;
 
 import com.adobe.acs.commons.util.BufferingResponse;
+import com.aemtask.core.config.CompanyNameConfig;
 import org.apache.sling.engine.EngineConstants;
 import org.osgi.framework.Constants;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.metatype.annotations.AttributeDefinition;
 import org.osgi.service.metatype.annotations.Designate;
 import org.osgi.service.metatype.annotations.ObjectClassDefinition;
@@ -22,53 +24,27 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.PrintWriter;
 
+/**
+ * Servlet filter component that replaces company names in content
+ */
 @Component(service = Filter.class, property = {
         Constants.SERVICE_DESCRIPTION + "= Filter incoming requests and replace company name in response",
         EngineConstants.SLING_FILTER_SCOPE + "=" + EngineConstants.FILTER_SCOPE_REQUEST,
         EngineConstants.SLING_FILTER_PATTERN + "=/content/we-retail/.*",
         Constants.SERVICE_RANKING + "=-700"
 })
-@Designate(ocd = CompanyReplaceFilter.Config.class)
 public class CompanyReplaceFilter implements Filter {
-
-    @ObjectClassDefinition(name = "Filter configuration")
-    public static @interface Config {
-        @AttributeDefinition(name = "source")
-        String source() default "";
-        @AttributeDefinition(name = "replacement")
-        String replacement() default "";
-    }
 
     private final Logger LOGGER = LoggerFactory.getLogger(getClass());
 
     private static final String CONTENT_TYPE_HTML = "html";
     private static final String CONTENT_TYPE_JSON = "json";
 
-    private String source;
-    private String replacement;
+    @Reference
+    private CompanyNameConfig companyNameConfig;
 
-    public String getSource() {
-        return source;
-    }
-
-    public void setSource(String source) {
-        this.source = source;
-    }
-
-    public String getReplacement() {
-        return replacement;
-    }
-
-    public void setReplacement(String replacement) {
-        this.replacement = replacement;
-    }
-
-    @Activate
-    protected void activate(final Config config) {
-        setSource(config.source());
-        LOGGER.info("configure: source='{}''", this.source);
-        setReplacement(config.replacement());
-        LOGGER.info("configure: replacement='{}''", this.replacement);
+    public CompanyNameConfig getCompanyNameConfig() {
+        return companyNameConfig;
     }
 
     @Override
@@ -84,7 +60,8 @@ public class CompanyReplaceFilter implements Filter {
         if (response.getContentType() != null
                 && (response.getContentType().contains(CONTENT_TYPE_HTML)
                 || response.getContentType().contains(CONTENT_TYPE_JSON))) {
-            String content = bufferingResponse.getContents().replaceAll(getSource(), getReplacement());
+            String content = bufferingResponse.getContents().replaceAll(getCompanyNameConfig().getSource(),
+                    getCompanyNameConfig().getReplacement());
             response.setContentLength(content.length());
             responseWriter.write(content);
         }
